refactor: model 하위 dto, entity 폴더 생성, 폴더 구조 변경
